src/requirements.txt
# src/requirements.txt

Django==5.2.4
gunicorn==22.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
django-extensions==3.2.3
whitenoise==6.7.0
django-admin-interface==0.28.8
django-colorfield==0.11.0
django-recaptcha==4.0.0
django-redis==5.4.0
redis==5.2.1
markdown2==2.4.13
django-ckeditor-5==0.2.13
pillow==10.4.0
pygments==2.17.2
django-debug-toolbar==4.3.0
django-crontab==0.7.1
